+++ jupyter/lib/modules/kernel.nix
@@ -19,5 +19,11 @@
       default = _: [];
       defaultText = lib.literalExpression ''_: []'';
     };
+
+    jupyterExtensions = lib.mkOption {
+      type = lib.types.listOf lib.types.package;
+      description = "Packages containing Jupyter extensions to install. Be careful with package sets to ensure compatibility";
+      default = [ ];
+    };
   };
 }